@@ -123,7 +123,7 @@
     "@fortawesome/fontawesome-free-solid": "^5.0.6",
     "@fortawesome/react-fontawesome": "^0.0.17",
     "@types/prop-types": "^15.5.2",
-    "@types/react": "15",
+    "@types/react": "^16.3.13",
     "@types/react-svg": "^2.2.0",
     "ReconnectingWebSocket": "git+https://github.com/bitshares/reconnecting-websocket",
     "alt": "^0.18.6",
@@ -143,6 +143,7 @@
     "counterpart": "^0.18.3",
     "create-hash": "^1.1.2",
     "create-hmac": "^1.1.4",
+    "create-react-class": "^15.6.3",
     "crypto-js": "^3.1.9-1",
     "deep-equal": "^1.0.1",
     "ecurve": "^1.0.2",
@@ -172,10 +173,10 @@
     "pubsub-js": "^1.6.0",
     "qrcode.react": "^0.7.1",
     "rc-queue-anim": "^1.2.2",
-    "react": "^15.6.1",
+    "react": "^16.3.2",
     "react-apollo": "^2.0.4",
     "react-clipboard.js": "^1.0.1",
-    "react-dom": "^15.6.1",
+    "react-dom": "^16.3.2",
     "react-highcharts": "^12.0.0",
     "react-hot-loader": "^3.0.0",
     "react-interpolate-component": "^0.10.0",
@@ -190,7 +191,7 @@
     "react-svg-inline": "^2.1.0",
     "react-tooltip": "^3.2.2",
     "react-transition-group": "^1.2.0",
-    "react-translate-component": "^0.13.2",
+    "react-translate-component": "^0.15.1",
     "rxjs": "^5.4.3",
     "secure-random": "^1.1.1",
     "tcomb": "2.5.2",
@@ -201,7 +202,7 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.92",
-    "@types/react-dom": "15",
+    "@types/react-dom": "^16.0.5",
     "@types/react-router": "3",
     "awesome-typescript-loader": "^3.2.3",
     "babel-core": "^6.21.0",
